Doxygen updates for PostureEstimator
